+++ src/Binlog/Collector/Model/BinlogHistoryModel.php
@@ -212,11 +212,11 @@ class BinlogHistoryModel extends BinlogHistoryBaseModel
         return intval($this->db->sqlCount('platform_universal_history_3_binlog', $where));
     }
 
-    public function getRecentEmptyGtidBinlogId(): ?string
+    public function getRecentEmptyGtidBinlogId(): int
     {
         $where = ['gtid' => ''];
 
-        return $this->db->sqlData(
+        return (int)$this->db->sqlData(
             'SELECT id FROM platform_universal_history_3_binlog WHERE ? ORDER BY id DESC ?',
             sqlWhere($where),
             sqlLimit(1)
@@ -251,14 +251,14 @@ class BinlogHistoryModel extends BinlogHistoryBaseModel
         );
     }
 
-    public function getEmptyGtidBinlogIdByLesserIdAndOffset(int $id, int $offset): ?string
+    public function getEmptyGtidBinlogIdByLesserIdAndOffset(int $id, int $offset): int
     {
         $where = [
             'gtid' => '',
             'id' => sqlLesser($id),
         ];
 
-        return $this->db->sqlData(
+        return (int)$this->db->sqlData(
             'SELECT id FROM platform_universal_history_3_binlog WHERE ? ORDER BY id DESC ?',
             sqlWhere($where),
             sqlLimit($offset, 1)