+++ src/Binlog/Collector/OnceBinlogHistoryService.php
@@ -114,7 +114,7 @@ class OnceBinlogHistoryService implements BinlogHistoryServiceInterface
         return $binlog_history_model->getEmptyGtidBinlogCount();
     }
 
-    public function getRecentEmptyGtidBinlogId(): ?string
+    public function getRecentEmptyGtidBinlogId(): int
     {
         $binlog_history_model = OnceBinlogHistoryModel::createBinlogHistoryWrite();
 
@@ -135,7 +135,7 @@ class OnceBinlogHistoryService implements BinlogHistoryServiceInterface
         return $binlog_history_model->getEmptyGtidBinlogDictsByLesserId($id, $limit);
     }
 
-    public function getEmptyGtidBinlogIdByLesserIdAndOffset(int $id, int $offset): ?string
+    public function getEmptyGtidBinlogIdByLesserIdAndOffset(int $id, int $offset): int
     {
         $binlog_history_model = OnceBinlogHistoryModel::createBinlogHistoryWrite();